giving-account/pyproject.toml
[tool.poetry]
name = "giving-account"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.74.0"
Authlib = "^0.15.5"
httpx = "^0.22.0"
itsdangerous = "^2.1.0"
uvicorn = "^0.17.5"
ulid-py = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
requests = "^2.27.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
